Rename java package
The `prometheus4cats.java` package name can conflict with the `java`
base package when `prometheus4cats._` is imported!
